fixed Framework issues and added EPS to core
